pyproject.toml
[project]
name = "exposeee-api"
version = "0.1.0"
description = "API interface for exposeee service"
authors = [
    {name = "joaosr",email = "eng.jmsoares@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (==5.1.7)",
    "django-cors-headers (==3.13.0)",
    "djangorestframework (>=3.14.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.2.2,<6.0.0)",
    "python-decouple (==3.6)",
    "dj-rest-auth (==2.2.5)",
    "django-allauth (==0.51.0)",
    "dj-database-url (==3.0.1)",
    "django-storages (==1.13.1)",
    "django-s3-storage (==0.13.10)",
    "psycopg2-binary (==2.9.10)",
    "django-test-without-migrations (==0.6)",
    "django-redis (==5.2.0)",
    "django-json-widget (==1.1.1)",
    "django-rq (>=3.0.0,<4.0.0)",
    "channels (==4.3.2)",
    "channels-redis (==4.3.0)",
    "sentry-sdk (>=2.24.0,<3.0.0)",
    "memba-match (>=0.1.8,<0.2.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "psycopg-pool (>=3.2.6,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "memba-match"
url = "https://test.pypi.org/simple/"
priority = "supplemental"


[tool.poetry.group.dev.dependencies]
pytest = "9.0.2"
pytest-django = "4.11.1"
pylint = "3.3.6"
flake8 = "7.1.2"
testcontainers = "^4.13.3"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
junit_family = "xunit1"